requirements.txt
-e git+https://github.com/pawanvirsingh/django-easy-pdf.git#egg=django-easy-pdf
adal==1.2.4
arabic-reshaper==2.1.0
asgiref==3.3.4
azure-common==1.1.25
beautifulsoup4==4.9.3
bokeh==2.2.3
boto3==1.17.18
cachetools==4.1.1
cairocffi==1.1.0
CairoSVG==2.5.1
certifi==2020.6.20
cffi==1.14.3
chardet==4.0.0
cryptography==3.3.2
cssselect2==0.4.1
cx-Oracle==7.1.3
defusedxml==0.6.0
Deprecated==1.2.10
django-bootstrap4==2.3.1
django-debug-toolbar==3.2.1
django-el-pagination==3.3.0
django-filter==21.1
django-phonenumber-field==5.0.0
django-storages==1.10.1
django-storages[azure]
django-tracking2==0.5.0
django-webpack-loader==0.7.0
Django==3.2.5
djangorestframework-csv==2.1.0
djangorestframework==3.12.2
factory-boy==3.1.0
Faker==4.14.0
future==0.18.2
google-api-core==1.23.0
google-auth==1.22.1
googleapis-common-protos==1.52.0
gunicorn==20.0.4
html2markdown==0.1.7
html2text==2020.1.16
html5lib==1.1
idna==2.10
isodate==0.6.0
Jinja2==2.11.3
Markdown==3.3.3
MarkupSafe==1.1.1
msrest==0.6.19
msrestazure==0.6.4
mysqlclient==1.4.6
numpy==1.19.2
oauthlib==3.1.0
opencensus-context==0.1.2
opencensus-ext-azure==1.0.5
opencensus-ext-django==0.7.2
opencensus==0.7.11
openpyxl==3.0.5
packaging==20.4
pandas==1.1.3
phonenumberslite==8.12.8
Pillow==8.3.2
protobuf==3.13.0
psutil==5.7.3
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20
PyGithub==1.53
PyJWT==1.7.1
pyparsing==2.4.7
PyPDF2==1.26.0
Pyphen==0.9.5
python-bidi==0.4.2
python-dateutil==2.8.1
python-decouple==3.3
python-docx==0.8.10
python-http-client==3.3.1
python-Levenshtein==0.12.0
pytz==2020.1
PyYAML==5.4
regex==2020.10.28
reportlab==3.5.54
requests-oauthlib==1.3.0
requests==2.25.1
rsa==4.7
selenium==3.141.0
sendgrid==6.4.7
shapely==1.7.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.2
starkbank-ecdsa==1.1.0
text-unidecode==1.3
textile==4.0.1
tinycss2==1.0.2
tornado==6.0.4
typing-extensions==3.7.4.3
unicodecsv==0.14.1
urllib3==1.26.5
WeasyPrint==51
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.12.1
xhtml2pdf==0.2.5
xlrd==2.0.1
XlsxWriter==1.3.7
Unidecode==1.2.0